# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Freshify API Documentation
  description: |-
    An API for creating and managing a household inventory and shopping list.
    The API is written in Java using the Spring Boot framework. The API uses a
    MySQL database. The API is documented using Swagger. The API is hosted using
    Apache.
    Some useful links:
    - [The GitLab repository](https://gitlab.stud.idi.ntnu.no/idatt2106-v23-10/smartmat-backend) 
  contact:
    email: vkbugge@hotmail.com
  license:
    name: MIT License
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /household:
    post:
      tags:
        - household
      summary: Create a household
      description: Creates a household
      operationId: createHousehold
      requestBody:
        description: The household to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Household'
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '409':
          description: Household already exists
  /household/{id}/users:
    get:
      tags:
        - household
      summary: Get household users
      description: Gets all the users that are member of a household
      operationId: getUsers
      parameters:
        - name: id
          in: path
          description: ID of household to get users from
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    user:
                      $ref: '#/components/schemas/UserFull'
                    userType:
                      $ref: '#/components/schemas/HouseholdUserType'
        '401':
          description: Access denied
        '404':
            description: Household not found
    put:
      tags:
        - household
      summary: Update household member role of user
      description: Updates the role of a user withing a given household. Can only be done by a superuser
      operationId: updateHouseholdMemberRole
      parameters:
        - name: id
          in: path
          description: ID of household to update a user type in
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated the type of user in the household
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHouseholdUserType"
      responses:
        '200':
          description: The updated household member
          content:
            schema:
              $ref: '#/components/schemas/HouseholdMember'
        '400':
          description: Illegal household member role/User does not belong to household
        '401':
          description: Access denied
        '404':
            description: Household not found/user not found
    post:
      tags:
        - household
      summary: Add user
      description: Adds a user to the household. Can only be done by a superuser
      operationId: addUser
      parameters:
        - name: id
          in: path
          description: ID of household to add a user to
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: User to add to the household
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  format: int64
      responses:
        '201':
          description: successful operation
        '409':
          description: User already exists in the household
        '401':
          description: Access denied
        '404':
            description: Household not found
...
  /household/{id}:
    delete:
      tags:
        - household
      summary: Delete a household
      description: Deletes a household. Can only be done by superusers.
      operationId: deleteHousehold
      parameters:
        - name: id
          in: path
          description: ID of household to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: successful operation
        '404':
          description: Household not found
        '401':
          description: Access denied
    get:
      tags:
        - household
      summary: Get household by id
      description: Returns a single household
      operationId: getHouseholdById
      parameters:
        - name: id
          in: path
          description: ID of household to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '404':
          description: Household not found
    put:
      tags:
          - household
      summary: Update household
      description: Updates the household's attributes. Can only be done by a superuser
      operationId: updateHousehold
      parameters:
        - name: id
          in: path
          description: ID of household to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: The household to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Household'
      responses:
          '200':
            description: successful operation
          '404':
            description: Household not found
          '401':
            description: Access denied
...
components:
  schemas:
...
    UserFull:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        email:
          type: string
          format: email
          example: example@example.org
        firstName:
          type: string
          example: John
...
    HouseholdUserType:
      type: string
      enum:
        - SUPERUSER
        - USER
    UpdateHouseholdUserType:
      type: object
      properties:
        userId:
          type: integer
          example: 123
        type:
          $ref: '#/components/schemas/HouseholdUserType'
    HouseholdMemberKey:
      type: object
      properties:
        householdId:
          type: integer
          example: 1
        userId:
          type: integer
          example: 1
    HouseholdMember:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/HouseholdMemberKey'
        user:
          $ref: '#/components/schemas/UserFull'
        type:
          $ref: '#/components/schemas/HouseholdUserType'
    Household:
      type: object
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: "My family"